chore: reorganise player files
