+++ resources/js/components/video/VideoPlayer.vue
@@ -26,20 +26,20 @@ import { FLAGS } from '@/config/featureFlags';
 
 import AudioSpectrographPanel from '@/components/video/audio/AudioSpectrographPanel.vue';
 import VideoControlWrapper from '@/components/video/VideoControlWrapper.vue';
-import VideoPopoverSlider from '@/components/video/VideoPopoverSlider.vue';
+import VideoPopoverSlider from '@/components/video/popover/VideoPopoverSlider.vue';
 import AudioSpectrograph from '@/components/video/audio/AudioSpectrograph.vue';
-import VideoPopoverItem from '@/components/video/VideoPopoverItem.vue';
+import VideoPopoverItem from '@/components/video/popover/VideoPopoverItem.vue';
 import PlayerSubtitles from '@/components/video/subtitles/PlayerSubtitles.vue';
-import VideoPartyPanel from '@/components/video/VideoPartyPanel.vue';
-import PlayerSkipIntro from '@/components/video/PlayerSkipIntro.vue';
+import VideoPartyPanel from '@/components/video/plugins/party/VideoPartyPanel.vue';
+import PlayerSkipIntro from '@/components/video/plugins/skip-intro/PlayerSkipIntro.vue';
 import PlayerBackdrop from '@/components/video/PlayerBackdrop.vue';
 import VideoTimeline from '@/components/video/VideoTimeline.vue';
 import VideoHeatmap from '@/components/video/VideoHeatmap.vue';
-import VideoPopover from '@/components/video/VideoPopover.vue';
-import VideoButton from '@/components/video/VideoButton.vue';
-import VideoSlider from '@/components/video/VideoSlider.vue';
-import VideoLyrics from '@/components/video/VideoLyrics.vue';
-import PlayerStats from '@/components/video/PlayerStats.vue';
+import VideoPopover from '@/components/video/popover/VideoPopover.vue';
+import VideoButton from '@/components/video/button/VideoButton.vue';
+import VideoSlider from '@/components/video/slider/VideoSlider.vue';
+import VideoLyrics from '@/components/video/plugins/lyrics/VideoLyrics.vue';
+import PlayerStats from '@/components/video/plugins/stats/PlayerStats.vue';
 
 import ProiconsPictureInPictureEnter from '~icons/proicons/picture-in-picture-enter';
 import ProiconsPictureInPictureExit from '~icons/proicons/picture-in-picture-exit';
