+++ resources/js/components/video/plugins/lyrics/VideoLyricItem.vue
