+++ resources/js/components/video/plugins/party/VideoPartyItem.vue
