+++ resources/js/components/video/plugins/stats/PlayerStats.vue
