+++ resources/js/components/video/popover/VideoPopoverItem.vue
