+++ resources/js/components/video/popover/VideoPopoverSlider.vue
