+++ resources/js/components/video/tooltip/VideoTooltipBase.vue
